requirements.txt
asgiref==3.5.1
attrs==24.2.0
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.3
colorama==0.4.4
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.8
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy-extensions==0.4.3
packaging==24.2
pathspec==0.9.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
sqlparse==0.4.2
swagger-spec-validator==3.0.4
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2022.1
uritemplate==4.1.1
zipp==3.20.2
